+++ spec/oxblood/commands_spec.rb
@@ -4,7 +4,7 @@ RSpec.describe Oxblood::Commands do
   specify do
     commands_groups = %w(
       Hashes HyperLogLog Strings Connection Server Keys Lists Scripting Sets
-      SortedSets Transactions
+      SortedSets Transactions Geo
     ).map { |g| described_class.const_get(g) }
 
     expect(described_class.included_modules).to match_array(commands_groups)